+++ assets/js/monaco/index.tsx
@@ -1,0 +1,20 @@
+import type { Monaco } from "@monaco-editor/react";
+
+export function setTheme(monaco: Monaco) {
+  monaco.editor.defineTheme('default', {
+    base: 'vs-dark',
+    inherit: true,
+    rules: [],
+    colors: {
+      'editor.foreground': '#E2E8F0',
+      'editor.background': '#334155', // slate-700
+      'editor.lineHighlightBackground': '#475569', // slate-600
+      'editor.selectionBackground': '#4f5b66',
+      'editorCursor.foreground': '#c0c5ce',
+      'editorWhitespace.foreground': '#65737e',
+      'editorIndentGuide.background': '#65737F',
+      'editorIndentGuide.activeBackground': '#FBC95A',
+    },
+  });
+  monaco.editor.setTheme('default');
+}