+++ lib/lightning_web/live/run_live/run_viewer_live.ex
@@ -184,10 +184,7 @@ defmodule LightningWeb.RunLive.RunViewerLive do
                   <div class="flex-1 grow inset-0 overflow-auto rounded-md">
                     <Viewers.step_dataclip_viewer
                       id={"step-input-#{@selected_step_id}"}
-                      class="overflow-auto h-full"
                       run_state={@run.result.state}
-                      stream={@streams.input_dataclip}
-                      stream_empty?={@input_dataclip_stream_empty?}
                       step={@selected_step}
                       dataclip={@input_dataclip}
                       input_or_output={:input}
@@ -234,9 +231,6 @@ defmodule LightningWeb.RunLive.RunViewerLive do
                   <div class="flex-1 grow inset-0 overflow-auto rounded-md">
                     <Viewers.step_dataclip_viewer
                       id={"step-output-#{@selected_step_id}"}
-                      class="overflow-auto h-full"
-                      stream={@streams.output_dataclip}
-                      stream_empty?={@output_dataclip_stream_empty?}
                       run_state={@run.result.state}
                       step={@selected_step}
                       dataclip={@output_dataclip}
@@ -278,12 +272,8 @@ defmodule LightningWeb.RunLive.RunViewerLive do
      )
      |> stream(:log_lines, [])
      |> assign(:log_lines_stream_empty?, true)
-     |> stream(:input_dataclip, [])
-     |> assign(:input_dataclip_stream_empty?, true)
-     |> assign(:input_dataclip, false)
-     |> stream(:output_dataclip, [])
-     |> assign(:output_dataclip_stream_empty?, true)
-     |> assign(:output_dataclip, false)
+     |> assign(:input_dataclip, nil)
+     |> assign(:output_dataclip, nil)
      |> assign(:run, AsyncResult.loading())
      |> assign(:log_lines, AsyncResult.loading())
      |> assign(