+++ lib/lightning_web/live/workflow_live/edit.ex
@@ -1033,7 +1033,7 @@ defmodule LightningWeb.WorkflowLive.Edit do
   end
 
   def handle_info(%DataclipUpdated{dataclip: dataclip}, socket) do
-    dataclip = Invocation.get_dataclip_details!(dataclip.id)
+    dataclip = Invocation.get_dataclip!(dataclip.id)
 
     {:noreply,
      assign_dataclips(socket, socket.assigns.selectable_dataclips, dataclip)}